+++ WMS/Scheduler.WebAPI/Controllers/ItemListsController.cs
@@ -124,7 +124,7 @@ namespace Ferretto.WMS.Scheduler.WebAPI.Controllers
                            Description = l.Description,
                            Priority = l.Priority,
                            ItemListStatus = (ItemListStatus)l.Status,
-                           ItemListType = (int)((ItemListType)l.ItemListType),
+                           ItemListType = (ItemListType)l.ItemListType,
                            ItemListItemsCount = l.ItemListRows.Sum(row => row.RequiredQuantity),
                            CreationDate = l.CreationDate,
                            Job = l.Job,
